% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaxes.R
\name{gaxes}
\alias{gaxes}
\title{Draws axes:
Called by grstereo}
\usage{
gaxes(r, nda)
}
\arguments{
\item{r}{Radius 0 < r <=1}

\item{nda}{0 (none), 1 (main only), 2 (main+110 axes)}
}
\description{
Draws axes:
Called by grstereo
}
