% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcadd.R
\name{gcadd}
\alias{gcadd}
\title{Draws circle on plane perpendicular to x3}
\usage{
gcadd(rad, lw, clr, r)
}
\arguments{
\item{rad}{Radius of circle to be drawn rad < r}

\item{lw}{Line width for circle}

\item{clr}{Colour of the circle for circle}

\item{r}{Radius of stereogram}
}
\description{
Draws circle on plane perpendicular to x3
}
