% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalcrz.R
\name{gcalcrz}
\alias{gcalcrz}
\title{Calculates radii of circle through
points with z + and z -}
\usage{
gcalcrz(xmat)
}
\arguments{
\item{xmat}{Matrix X}
}
\value{
radii Radii1 for z>0 and  Radii2 for z < 0
}
\description{
Calculates radii of circle through
points with z + and z -
}
