% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gccalc.R
\name{gccalc}
\alias{gccalc}
\title{Centre and radius of circle thru 3 points on the
equatorial plane specified in xymat}
\usage{
gccalc(xymat)
}
\arguments{
\item{xymat}{}
}
\value{
c(gc,ra) gc:coordinates of centre. ra:radius
}
\description{
Centre and radius of circle thru 3 points on the
equatorial plane specified in xymat
}
