% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcircle.R
\name{gcircle}
\alias{gcircle}
\title{Draws circle of stereogram and places axis labels:
Called by grstereo}
\usage{
gcircle(r, nbb)
}
\arguments{
\item{r}{Radius of circle 0 < r <= 1}

\item{nbb}{Labels 0 (x1,x2,x3), 1 (110 added)}
}
\description{
Draws circle of stereogram and places axis labels:
Called by grstereo
}
