% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmeridian.R
\name{gmeridian}
\alias{gmeridian}
\title{Draws planes through NS poles: medians/longitudes}
\usage{
gmeridian(d, lw, clr, r)
}
\arguments{
\item{d}{Degree of rotation}

\item{lw}{Line width}

\item{clr}{Colour}

\item{R}{Radius of stereogram}
}
\description{
Draws planes through NS poles: medians/longitudes
}
