% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.R
\name{gplot}
\alias{gplot}
\title{Plot points with colour by converting to
stereogram coordinates bigx bigy}
\usage{
gplot(xmat, cx, clr, r)
}
\arguments{
\item{xmat}{Matrix of 3 columns for points}

\item{cx}{Size of points}

\item{clr}{Colour of points}

\item{r.}{Radius of stereogram}
}
\description{
Plot points with colour by converting to
stereogram coordinates bigx bigy
}
