% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great.R
\name{greatc}
\alias{greatc}
\title{Draws great circles. Calls gccalc}
\usage{
greatc(r, ngc)
}
\arguments{
\item{r}{Radius for drawing the circle}

\item{ngc}{Axes: 0 (none), 1 (main axis ), 2 (main + 110 axes)}
}
\description{
Draws great circles. Calls gccalc
}
