% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grstereo.R
\name{grstereo}
\alias{grstereo}
\title{Draws circle, axes and labels of a stereogram
Calls functions gaxes and greatc}
\usage{
grstereo(r, nda, ngc, nbb)
}
\arguments{
\item{r}{Radius of stereogram}

\item{nda}{Axes: 0 (none), 1(x,y), 2 (x,y + 110)}

\item{ngc}{Great circles: 0 (none), 1 (x,y), 2 (x,y+110)}

\item{nbb}{Labels 0 (x1,x2,x3), 1 (110 added)}
}
\description{
Draws circle, axes and labels of a stereogram
Calls functions gaxes and greatc
}
\examples{
gstereo(1,1,0,1)

}
