% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtcirc111.R
\name{gtcirc111}
\alias{gtcirc111}
\title{Plot four circles round [111] axes
Calls function gccalc}
\usage{
gtcirc111(xc, clr, lw, r)
}
\arguments{
\item{xc}{Coordinates of circle around (111)}

\item{clr}{Colour of circles}

\item{lw}{Line width for circles}

\item{r}{Radius of stereogram}
}
\description{
Plot four circles round [111] axes
Calls function gccalc
}
