% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grotatez.R
\name{rmatz}
\alias{rmatz}
\title{Matrix to rotate points on sphere around z-axis}
\usage{
rmatz(theta)
}
\arguments{
\item{theta}{Angle of rotation}
}
\value{
rmatz Rotation matrix
}
\description{
Matrix to rotate points on sphere around z-axis
}
