% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwvec9.R
\name{wvvector}
\alias{wvvector}
\title{Weights for Euclidean 9-design on
Yamamoto, Hirao, Sawa (2019)}
\usage{
wvvector(A, r2)
}
\arguments{
\item{A}{Parameter for calculating weights
a=square root of A}

\item{r2}{Radius of concentric circle}
}
\value{
wvec Vector of weights
             (w2,w3,wint,v1,v2,v3)
}
\description{
Weights for Euclidean 9-design on
Yamamoto, Hirao, Sawa (2019)
}
